requirement.txt
certifi==2019.9.11
chardet==3.0.4
gitdb2==2.0.5
GitPython==3.0.2
idna==2.8
img2pdf==0.3.3
Pillow==6.1.0
requests==2.22.0
selenium==3.141.0
smmap2==2.0.5
urllib3==1.25.4
